Move DemosLite.svelte to DemosLite/DemosLite.svelte
