+++ js/_website/src/lib/components/DemosLite/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DemosLite.svelte";
